fix(lexer): small error fixing and minor improvement for minor error handling.
To check the end of a literal ("text"), the program read in an
out-of-bound area if the literal was not complete. This trigger a panic.

In the other tokens, the `current_byte` was updated to the offset. The
correct logic was the addition of the offset to the current position.

- Added bound-check in the literals
- Correction of all current_byte manipulations.

Additiaonaly,  hyphen and under-hyphen were no considered in the term names. They were added
to  the lexer rules.

Finnaly, for error handling in the future, the  token position in file was added to the Token struct.