feat: added structural parsing of tokens.
This commit implements of the structural parsing of the
lexer tokens.

It contains the parsing of the prefixes that are given by:
```
@prefix xx: <example.com>.
@base <example.com>.
""
Prefix & base tests were added to check possible cases. The test are not
very exhaustive and they only check the most valid cases and their presence of invalid tokens.

Structure parsing of complex objects and nested nodes. This code implements new errors to represent better error messages to improve the mapping learning.